backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
google-cloud-storage==2.14.0
python-multipart==0.0.9
sentence-transformers==2.6.1
scikit-learn==1.4.1.post1
PyMuPDF==1.23.22
pandas==2.2.1
numpy==1.26.4
